refactor(xtask): extract per-token parity ladder helpers
